% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acid.df}
\alias{acid.df}
\title{Age estimation from aspartic acid concentration}
\format{A data.frame with 37 rows and 3 columns:
\describe{
\item{age}{Age in years.}
\item{period}{Period of tooth, modern or victorian.}
\item{aspartic}{Percentage of D-aspartic acid.}
}}
\source{
Gillard, R.D., Hardman, S.M., Pollard, A.M., Sutton, P.A. and
Whittaker, D.K. (1991) 'Determinations of age at death in the archaeological populations
using the D/L ratio of aspartic acid in dental collagen' in Archaeometry 90,
eds. Pernicka, E. and Wagner, G.A., p.637-644, Birkhauser Verlag, Berlin.
}
\usage{
acid.df
}
\description{
Aspartic acid data for modern upper and lower first pre-molars: taken from
Gillard et al 1991
}
\keyword{datasets}
