% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{activity.df}
\alias{activity.df}
\title{Energy requirements for different activities}
\format{
A data.frame with 24 rows and 3 columns:
\describe{
\item{subject}{a subject ID.}
\item{activity}{running, walking, riding.}
\item{energy}{energy expended during activity, in kilocalories (Cal)}
}
}
\source{
Milton, J. S. (1992). Statistical Methods in the Biological and Health Sciences
2nd Edition, McGraw-Hill, New York, p. 316--319.
}
\usage{
activity.df
}
\description{
An experiment was conducted to compare the energy requirements of three
physical activities: running, walking and bicycle riding. Eight subjects were
asked to run, walk and bicycle a measured distance, and the number of
kilocalories expended per kilometre was measured for each subject during
each activity. The activities are run in random order with time for recovery
between activities. Each activity was monitored exactly once for each
individual.
}
\keyword{datasets}
