% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleSort.R
\name{bubbleSort}
\alias{bubbleSort}
\title{Bubble sort}
\usage{
bubbleSort(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
the vector \code{x} sorted into ascending order
}
\description{
Sorts the vector x into ascending order using a very inefficient bubble sort algorithm
}
\examples{
set.seed(123)
x = rnorm(10)
bubbleSort(x)
}
