% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calculus.df}
\alias{calculus.df}
\title{Calculus marks}
\format{
A data.frame with 147 rows and two columns:
\describe{
\item{coursework}{final coursework mark out of 100.}
\item{examination}{final examination mark out of 100.}
}
}
\source{
George Moran, Department of Mathematics and Statistics at Lancaster.
}
\usage{
calculus.df
}
\description{
Calculus marks from the 2012 first year calculus course from the
Department of Mathematics and Statistics at Lancaster
University.
}
\keyword{datasets}
