% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{car.prices.df}
\alias{car.prices.df}
\title{Car listings from trademe}
\format{
An object of class \code{data.frame} with 3618 rows and 13 columns.
}
\usage{
car.prices.df
}
\description{
This data set consists of 3,618 listings scraped from the New Zealand website \href{http://trademe.co.nz}{tradee}.
trademe is similar to \href{ebay.com}{ebay} in that is an online auction site which allows sellers to list new
and used goods for sale. Goods may be purchased via auction, or outright if the seller has enabled that option. Many
New Zealanders, including commericial car dealers, use the website to buy and sell cars. The listings gathered consist
are mostly for Mazda 3 and Toyota Corolla vehicles, along with imported vehicles which may be the same car, but with
different badging.
}
\details{
@format A data.frame with 3,618 rows and 14 columns:
 \describe{
 \item{obs}{The observation number, from 1 to 3618.}
 \item{title}{The listing title - basically the make and model of the car.}
 \item{year}{The year of manufacture of the vehicle.}
 \item{age}{The age of the vehicle as of 2013 (when this data was collected). So a car manufactured in 2009 would have an
 age of 4, for example.}
 \item{price}{The asking price, in NZD.}
 \item{km}{The number of kilometres on the odometer---i.e. the "mileage."}
 \item{cc}{The displacement of the engine in cubic centimetres.}
 \item{fuel}{The fuel used by  the vehicle: either Petrol (gasoline) or Diesel.}
 \item{doors}{The number of doors in the car. Note 3 and 5 door cars are hatchbacks.}
 \item{list.color}{The colour of the car given in the listing.}
 \item{simple.color}{An attempt to standardise the colour to a reduced category. For example sky blue, and light blue would
 both get transformed to blue.}
 \item{make}{The manufacturer of the car: either Mazda or Toyota}
 }
}
\keyword{datasets}
