% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carbon.df}
\alias{carbon.df}
\title{Carbon isotopes in trees}
\format{A data.frame with 200 rows and 4 columns:
\describe{
  \item{year}{}
  \item{iso}{}
  \item{temp}{}
}}
\usage{
carbon.df
}
\description{
These observations were made by Robertson et. al.
They are the mean delta 13 C compositions of several
individual trees from two locations in Central England
mean temperatures from the CET are also given
}
\keyword{datasets}
