% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_surv.df}
\alias{cell_surv.df}
\title{Cell survival data}
\format{An object of class \code{data.frame} with 27 rows and 2 columns.}
\usage{
cell_surv.df
}
\description{
The data comes from an experiment to measure the mortality of cancer cells
under radiation under taken in the Department of Radiology, University of Cape
Town. Four hundred cells were placed on a dish, and three dishes were
irradiated at a time, or occasion. After the cells were irradiated, the
surviving cells were counted. Since cells would also die naturally, dishes
with cells were put into the radiation chamber without being irradiated, to
establish the natural mortality. These data gives only these zero-dose data.
these data are from ozDASL
}
\keyword{datasets}
