% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chocolate.df}
\alias{chocolate.df}
\title{Energy and fat in chocolate bars}
\format{A data.frame with 16 rows and 2columns:
\describe{
\item{energy}{energy, in Calories = kilocalories}
\item{fat}{fat content, in grams}
}}
\source{
Source is unknown, but we would be happy to give credit if
someone tells us.
}
\usage{
chocolate.df
}
\description{
The amount of fat (g) and energy (Cal) in 16 chocolate bars.
Source is unknown, but we would be happy to give credit if
someone tells us.
}
\keyword{datasets}
