% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fingerprints.df}
\alias{fingerprints.df}
\title{Height, weight and fingerprint measurements collected from 200 participants}
\format{
a data.frame with 200 rows and 11 columns:
\describe{
\item{number}{participant number}
\item{gender}{self-declared gender of participant \code{female} or \code{male}}
\item{age}{age in years}
\item{hand}{dominant hand \code{left} or \code{right}}
\item{height}{height in centimetres, average of three measurements}
\item{weight}{weight in kilograms, average of three measurements}
\item{temp}{fingerprint temperature in degrees Celius}
\item{fpheight}{fingerprint height in millimetres}
\item{width}{fingerprint width in millimetres}
\item{area}{fingerprint area in squared millimetres}
\item{circumference}{fingerprint circumference in millimetres}
}
}
\source{
McMurchie, Beth; Torrens, George; Kelly, Paul (2019).
Height, weight and fingerprint measurements collected from 200 participants.
Loughborough University. [Dataset](https://doi.org/10.17028/rd.lboro.7539206.v1)
}
\usage{
fingerprints.df
}
\description{
This dataset contains the height, weight and 4 fingerprint measurements
(length, width, area and circumference), collected from 200 participants.
This data was collected with the intention of performing regression analysis
to asses whether a significant relationship exists between fingerprint size
and physical stature.
}
\keyword{datasets}
