% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModel.R
\name{getModel}
\alias{getModel}
\title{Get a JAGS model file}
\usage{
getModel(modelID)
}
\arguments{
\item{modelID}{a string containing a valid model ID}
}
\value{
a string containing the model. The intention is that this
can be written to disk.
}
\description{
This function provides an easy way for readers to get the JAGS model
files used in the book. The \code{modelID} is the 4-5 character identifier
used in the book. For example to get 'model-001.bugs.R', you would use
\code{getModel("001")}.
}
\examples{
getModel("001")
}
