% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gustafson.df}
\alias{gustafson.df}
\title{Age estimation based on changes in dental characteristics}
\format{a data.frame with 759 rows and 10 columns:
\describe{
\item{sex}{sex of subject, female or male.}
\item{age}{age, in years.}
\item{quadrant}{location in mouth of tooth}
\item{tooth}{tooth identifier}
\item{attrition}{}
\item{recession}{}
\item{dentine}{qualitative assessment of remaining dentine}
}}
\usage{
gustafson.df
}
\description{
Age estimation based on changes in dental characteristics
}
\keyword{datasets}
