% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{insulation.df}
\alias{insulation.df}
\title{Impact strength of insulation cuts in foot-pounds.}
\format{a data.frame with 100 rows and 3 colums:
\describe{
\item{Lot}{Lot of insulating material}
\item{Cut}{Lengthwise (Length) or crosswise (Cross)}
\item{Strength}{Impact strength, in foot-pounds (ft-lb)}
}}
\source{
Ostle, B. (1963). Statistics in Research: Basic Concepts and Techniques for Research. 
Ames, Iowa. Iowa State University Press.
}
\usage{
insulation.df
}
\description{
Impact strength of insulation cuts in foot-pounds.
}
\keyword{datasets}
