% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaggR-package.R
\docType{package}
\name{jaggR}
\alias{jaggR}
\title{jaggR: Supporting files and functions for the book Bayesian Modelling with JAGS}
\description{
A set of functions used in teaching STATS 201/208 Data Analysis at the University of Auckland. The functions are designed to make parts of R more accessible to a large undergraduate population who are mostly not statistics majors.
}
\author{
James Curran, David Lucy
}
