% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lightspeed.df}
\alias{lightspeed.df}
\title{Michelson's speed of light data}
\format{
a data.frame with 43 rows and 2 columns:
\describe{
\item{speed}{The scaled speed of light measured in a single experiment. The
scaling is the measurement minus 299,000 km/s. E.g. the first entry in the
data.frame is 850, which is 299,850 km/s.}
\item{year}{The year in which the experiment was conducted, either 1879 or 1882.}
}
}
\source{
Stigler, S. M. (1977), "Do robust estimators work with real data?",
The Annals of Statistics 5:1055-1098.
}
\usage{
lightspeed.df
}
\description{
Michelson's speed of light data
}
\keyword{datasets}
