% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mortality.df}
\alias{mortality.df}
\title{Mortality rates for different species}
\format{
a data.frame with 2117 rows and 4 columns:
\describe{
\item{group}{a factor indicating which one of the six taxonimic groups the observation
belongs to: bird, fish, invertebrate, mammal, multicellular plant, and phytoplankton.}
\item{species}{the species of the observation.}
\item{mass}{the body mass in grams (g).}
\item{mortality}{the mortality rate.}
\item{temp}{the average body temperature in degrees Celcius.}
\item{E}{average activation energy of heterotrophic respiration in animals (0.65 eV)
or photosynthesis in plants (0.32 eV).}
\item{mort.corrected}{mortality corrected by a Boltzmann-Arrhenius factor, specifically, divided by exp(-E/k * (1 / T - 1 / T20)),
where k is Boltzmann constant 8.62 x 10^-5, T20 is 20 degrees Celcius in degrees Kelvin, i.e. 293, and T is average
body temperature \code{temp} in degrees Kelvin.}
}
}
\source{
McCoy, M.W. and Gillooly, J.F. (2008), Predicting natural mortality
rates of plants and animals. Ecology Letters, 11: 710-716.
\href{https://doi-org.ezproxy.auckland.ac.nz/10.1111/j.1461-0248.2008.01190.x}{https://doi-org.ezproxy.auckland.ac.nz/10.1111/j.1461-0248.2008.01190.x}
}
\usage{
mortality.df
}
\description{
Ecologists Michael McCoy and James Gillooly were interested in predicting
mortality rates for different species based on a number of variables including
body mass, temperature. In their paper (McCoy and Gillooly, 2008) they explore
the hypothesis that the natural logarithm of temperature‐corrected mortality rate
should be a linear function of the natural logarithm of body mass. The
temperature-corrected mortality rate is based upon previous work which draws
on results from biology, biochemistry, and thermodynamics. Users are encouraged
to read the original source for a deeper explanation.
}
\keyword{datasets}
