% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plane.df}
\alias{plane.df}
\title{Distance travelled by paper planes}
\format{A data.frame with 16 rows and 5 columns:
\describe{
  \item{distance}{Distance travelled in mm.}
  \item{paper}{Paper weight in grams per square metre (gsm), either 80 gsm or 50 gsm.}
  \item{angle}{Angle of launch, horizontal or 45 degrees.}
  \item{design}{Design of the plane, either high performance or simple.}
}}
\source{
Mackisack, M. S. (1994). What is the use of experiments conducted by
statistics students? Journal of Statistics Education, 2, no 1.
}
\usage{
plane.df
}
\description{
A group from Queensland University of Technology conducted an experiment
where they recorded the distance flown by paper aeroplanes. The experimenters
used a sealed corridor at the University, and controlled the design of the
aeroplane, the weight of the paper from which each aeroplace was constructed,
and the angle of incidence at launch for each paper plane. The data and
further notes for this experiment can be found at
http://www.statsci.org/data/oz/planes.html.
}
\keyword{datasets}
