% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{radiation.df}
\alias{radiation.df}
\title{from the Commission facility in Hanford, Washington. One of the major safety
problems encountered there has been the storage of radioactive wastes. Over
the years, significant quantities of these substances - including strontium
90 and cesium 137 - have leaked from their open-pit storage areas into the
nearby Columbia River, which flows along the Washington-Oregon border, and
eventually empties into the Pacific Ocean.}
\format{An object of class \code{data.frame} with 9 rows and 3 columns.}
\source{
Fadeley, R. C. (1965). Oregon malignancy pattern physiographically
related to Hanford, Washington, Radioisotope Storage. Journal of
Environmental Health 27, 883-897.
}
\usage{
radiation.df
}
\description{
To measure the health consequences of this contamination, an index of
exposure was calculated for each of the nine Oregon counties having frontage
on either the Columbia River or the Pacific Ocean. This particular index was
based on several factors, including the county's stream distance from
Hanford and the average distance of its population from any water frontage.
As a covariate, the cancer mortality rate was determined for each of these
same counties.
The data give the index of exposure and the cancer mortality rate during
1959-1964 for the nine Oregon counties affected. Higher index values
represent higher levels of contamination.
}
\keyword{datasets}
