% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ratmaze.df}
\alias{ratmaze.df}
\title{Times taken for a rat to navigate through a maze}
\format{A data.frame with 135 rows and 4 columns:
\describe{
\item{subject}{An ID for each rat}
\item{treatment}{The treatment adminstered to the subject: control/none, thouiracil, thyroxin.}
\item{test}{A maze number.}
\item{time}{time, in seconds taken for the rat to navigate the maze.}
}}
\usage{
ratmaze.df
}
\description{
Times taken for a rat to navigate through a maze
}
\keyword{datasets}
