% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdt.df}
\alias{rdt.df}
\title{Age estimation by root dentine translucency}
\format{A data.frame with 71 rows and 3 columns:
\describe{
\item{age}{Age of subject, in years}
\item{sex}{Sex of subject, female or male}
\item{rdt}{root dentine translucency}
}}
\source{
Lucy, D., Aykroyd, R.G., Pollard, A.M. and Solheim (1996), T.,"A
Bayesian approach to adult human age estimation from dental observations
by Johanson's age changes", Journal of Forensic Sciences 41(2):189-194.
}
\usage{
rdt.df
}
\description{
Root dentine translucency is, in humans, an age related physiological
feature. In the dentine of teeth in adult humans the
tubecular microstructures fill with a highly crystalline substance making
them become nearly invisible when looked at in normal light. This process
starts from the apical foramen in early adulthood, and progresses up the
tooth into advanced old age. Solheim (Lucy et al., 1996) collected data
on age, root dentine translucency for 71 maxillary second incisors from a
Norweigian population. The sex of each individual was also noted.
}
\keyword{datasets}
