% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.mcmc.list.R
\name{reorder.mcmc.list}
\alias{reorder.mcmc.list}
\title{Reorder the columns of mcmc objects in an mcmc.list}
\usage{
\method{reorder}{mcmc.list}(x, variable.names, ...)
}
\arguments{
\item{x}{an object of type \code{mcmc.list}}

\item{variable.names}{a vector of variable names in user order.}

\item{...}{other arguments. Currently ignored.}
}
\value{
an object of type \code{mcmc.list}
}
\description{
Reorders the output from \code{\link{rjags}{coda.samples}} to match
the preferred order of the user. The function will stop if one or more of the
specified variable names does not match the variable names in the first
\code{mcmc} object of \code{x}.
}
