% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shotgun.df}
\alias{shotgun.df}
\title{Shotgun range data
In order to test the validity of range-of-fire estimates obtained by
the application of regression analysis to shotgun pellet patterns, a
blind study was conducted in which questioned pellet patterns were
fired at randomly selected ranges between 3.0 and 15.2 m (10 and 50 ft)
with two different 12-gauge shotguns. each firing a different type
of buckshot cartridge. Test firings at known ranges were also
conducted with the same weapons and ammunition.}
\format{
A data frame with 70 observations on 4 variables.
\describe{
\item{range}{The range in feet of the firing.}
\item{gun}{The model of shotgun used in the experiment.}
\item{expt}{A factor recording whether the data was to be used for building/training the model, or testing it.}
\item{area}{The area of the smallest rectangle that would enclose the pellet pattern.}
 }
}
\source{
Rowe, W.F. and Hanson, S.R. (1985) Range-of-fire estimates from regression
analysis applied to the spreads of shotgun pellet patterns: Results of a blind study,
Forensic Science International, 28(3-4): 239-250.
}
\usage{
shotgun.df
}
\description{
Shotgun range data
In order to test the validity of range-of-fire estimates obtained by
the application of regression analysis to shotgun pellet patterns, a
blind study was conducted in which questioned pellet patterns were
fired at randomly selected ranges between 3.0 and 15.2 m (10 and 50 ft)
with two different 12-gauge shotguns. each firing a different type
of buckshot cartridge. Test firings at known ranges were also
conducted with the same weapons and ammunition.
}
\keyword{datasets}
