% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{terriers.df}
\alias{terriers.df}
\title{Simulated weights of difference breeds of terriers}
\format{A data.frame with 30 rows and 2 columns.
\describe{
\item{weight}{Weight of dog in kg.}
\item{breed}{Breed, either Skye, Manchester or Norwich.}
}}
\usage{
terriers.df
}
\description{
Simulated samples of weights from English terrier breeds
with the parameter values for the means for the simulation
taken from http://www.dogsindepth.com. The variances are
assumed to be constant.
}
\keyword{datasets}
