% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdt.df}
\alias{rdt.df}
\title{Age estimation by root dentine translucency}
\format{A data.frame with 71 rows and 3 columns:
\describe{
\item{age}{Age of subject, in years}
\item{sex}{Sex of subject, female or male}
\item{rdt}{root dentine translucency}
}}
\usage{
rdt.df
}
\description{
Age estimation by root dentine translucency
}
\keyword{datasets}
