% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPlotPrefs.R
\name{setPlotPrefs}
\alias{setPlotPrefs}
\title{Set Plotting Preferences}
\usage{
setPlotPrefs(mar = c(3, 4, 1, 1), cex = 1, oma = c(0, 0, 0, 0),
  tcl = -0.35, mgp = c(1.5, 0.5, 0), las = 1, cex.lab = 1,
  font.lab = 1, lwd = 1, on.graph.line = 3, shading.density = 8,
  arrow.length = 0.1, on.graph.cex = 1, margin.cex = 1.2, ...)
}
\arguments{
\item{mar}{plot margings}

\item{cex}{character expansion factor}

\item{oma}{outer margins}

\item{tcl}{tick length}

\item{mgp}{somethen}

\item{las}{text rotation on axes}

\item{cex.lab}{plot labels cex}

\item{font.lab}{font of plot labels}

\item{lwd}{line width}

\item{on.graph.line}{no idea}

\item{shading.density}{shading density}

\item{arrow.length}{arrow head length}

\item{on.graph.cex}{character expansion for text on graphs}

\item{margin.cex}{character expansion for text for margins}

\item{...}{other arguments to be passed to par}
}
\value{
the previous par settings so that they can be restored
}
\description{
Set Plotting Preferences
}
